small fixes

+++ modules/StockPriece/include/S32_NumMethodsWZ.h
@@ -12,7 +12,7 @@ public:
 	double BurragePlatenStep(double S, double dt, double dw, double dz);
 	double Taylor2Step(double S, double dt, double dw, double dz);
 
-	void SimulateWandZProcesses(VSLStreamStatePtr stream, int nSteps, double Time, double *buffer);
+	void SimulateWandZProcesses(VSLStreamStatePtr stream, int nSteps, double Time, float *buffer);
 	void SimulateStockPrices(Step _step, VSLStreamStatePtr stream, int nPaths, int nSteps, double Time, double *Error);
 
 	void Execute(Step _step, char* FileName);
